requirement.txt
websockets==11.0.3
boto3
paho-mqtt
